requirements.txt
flask
flask-cors
flask-jwt-extended
flask-mail
pymysql
python-dotenv
reportlab


